backend/requirements.txt
# SlabSense Backend Dependencies
# Python 3.10+ required

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image Processing
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# OCR (optional - for card identification)
# pytesseract>=0.3.10

# Async Support
aiofiles>=23.2.0

# CORS (included with starlette via fastapi)
# starlette already included via fastapi

# Environment Variables
python-dotenv>=1.0.0

# HTTP Client (for Supabase verification)
httpx>=0.26.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Production (uncomment for deployment)
# gunicorn>=21.2.0
# redis>=5.0.0
# celery>=5.3.0
